refactor(datamodel): delete ast_to_dml::helpers
It was just EnvFunction, which is only used in configuration.rs, so it
moves there and becomes private.
