--- libs/datamodel/core/src/transform/helpers/mod.rs
@@ -1,3 +1,0 @@
-mod env_function;
-
-pub(crate) use env_function::EnvFunction;
